Revert "Add i18n"
