--- next-i18next.config.js
@@ -1,8 +1,0 @@
-/** @type {import('next-i18next').UserConfig} */
-module.exports = {
-  debug: process.env.NODE_ENV === 'development',
-  i18n: {
-    defaultLocale: 'en',
-    locales: ['en', 'ja'],
-  },
-};
