+++ next.config.js
@@ -1,9 +1,8 @@
-const withExportImages = require('next-export-optimize-images');
+const withExportImages = require("next-export-optimize-images");
 
 const nextConfig = withExportImages({
   reactStrictMode: true,
   swcMinify: true,
-  trailingSlash: true,
 });
 
 module.exports = nextConfig;
