+++ package.json
@@ -15,13 +15,9 @@
     "@emotion/styled": "^11.10.4",
     "@octokit/rest": "^19.0.4",
     "framer-motion": "^6.5.1",
-    "i18next": "^23.4.5",
     "next": "12.3.1",
-    "next-i18next": "^14.0.0",
-    "next-language-detector": "^1.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-i18next": "^13.2.0",
     "react-icons": "^4.4.0"
   },
   "devDependencies": {
