--- pages/[...slug].js
@@ -1,2 +1,0 @@
-import { Redirect } from 'lib/redirect';
-export default Redirect;
