+++ pages/_document.js
@@ -1,13 +1,10 @@
 import { ColorModeScript } from '@chakra-ui/react';
 import NextDocument, { Html, Head, Main, NextScript } from 'next/document';
-import i18nextConfig from '../next-i18next.config';
 
 export default class Document extends NextDocument {
   render() {
-    const currentLocale = this.props.__NEXT_DATA__.query.locale || i18nextConfig.i18n.defaultLocale;
-
     return (
-      <Html lang={currentLocale}>
+      <Html>
         <Head />
         <body>
           <ColorModeScript />
