--- public/locales/en/common.json
@@ -1,3 +1,0 @@
-{
-  "description": "Floorp is built on Firefox and was built in Japan and is a new browser with excellent privacy & flexibility."
-}
