--- public/locales/en/download-page.json
@@ -1,3 +1,0 @@
-{
-  "download-alert": "If you download Floorp you agree to the <h1>Pricacy Policy</h1>."
-}
