--- public/locales/en/landing-page.json
@@ -1,46 +1,0 @@
-{
-  "feature": "Features",
-  "features": [
-    {
-      "title": "Strong Tracking Protection",
-      "description": "Floorp includes a strong tracking blocker by default, protecting users from a variety of malicious tracking lurking on the web. In addition, there is fingerprinting protection."
-    },
-    {
-      "title": "Flexible Layout",
-      "description": "Floorp's layout customization is unlimited. Move the tab bar to the bottom of the window, hide the title bar, and more. Create your own Floorp."
-    },
-    {
-      "title": "Switchable Design",
-      "description": "In addition to the usual Firefox heme customization, allow switch between five different designs for Floorp interface and OS specific design."
-    },
-    {
-      "title": "Regular Updates",
-      "description": "Floorp is based on Firefox ESR. Floorp will be updated every 4 weeks, and security updates will be provided before Firefox release."
-    },
-    {
-      "title": "No User Tracking",
-      "description": "We don't get personal information from users. We don't track users. We don't sell user data. We don't have any relationship with any advertising company."
-    },
-    {
-      "title": "Completely Open Source",
-      "description": "The source code for Floorp is fully public, allowing anyone to view the source code and build Floorp. It is not only the browser itself that is open to the public. The build environment is also open source."
-    }
-  ],
-  "columns": [
-    {
-      "title": "Dual Sidebar",
-      "description": "Are you satisfied with only one sidebar? Want to open your website in the sidebar? Floorp has a built-in sidebar that can display a webpanel and browser management tools for a comfortable browsing experience. <br>A sidebar for busy people! Check everything from bookmarks and history to websites!"
-    },
-    {
-      "title": "Flexible Toolbar & Tab Bar",
-      "description": "Tree Style Tabs and use it as vertical tabs or in combination. Customize your bookmarks bar and more. It is such a browser that you can reach the itch. <br>We deliver the best customization for everyone from customization beginners to experts, such as custom CSS for those who want to be particular and basic themes!"
-    },
-    {
-      "title": "Think about the USER & the WEB",
-      "description": "Floorp does not track users and only blocks malicious tracking on websites. With Floorp you can build an internet that benefits both websites and users.<br> Blocking only harmful trackers by listening to the opinions of both website owners who make a living from advertising and users who want to protect their privacy, why not use Floorp in finding a better internet?"
-    }
-  ],
-  "latest-article": "Latest article",
-  "see-more": "See more",
-  "download-floorp": "Get latest version of Floorp Browser"
-}
