--- public/locales/ja/common.json
@@ -1,3 +1,0 @@
-{
-  "description": "Floorp は Firefox ベースで作られており、日本で誕生し、プライバシーに優れた新しいブラウザです。"
-}
