--- public/locales/ja/download-page.json
@@ -1,3 +1,0 @@
-{
-  "download-alert": "ダウンロードすることで<1>プライバシーポリシー</1>に同意したものとします。"
-}
