--- public/locales/ja/landing-page.json
@@ -1,46 +1,0 @@
-{
-  "feature": "特徴",
-  "features": [
-    {
-      "title": "強力なトラッカー保護",
-      "description": "Floorp は既定で強力な追跡遮断機能が含まれており、ウェブに潜む様々な悪意あるトラッキングからユーザーを守ります。更に、指紋採取対策も。"
-    },
-    {
-      "title": "幅広いカスタマイズ",
-      "description": "Floorp のレイアウトカスタマイズは無制限。タブバーの位置をウインドウの一番下に移動したりタイトルバーを非表示にしたり。あなただけの Floorp を。"
-    },
-    {
-      "title": "切り替え可能なデザイン",
-      "description": "Firefox の普通のテーマによるカスタマイズに加え、５つのデザインからブラウザーのインターフェイスを切り替えられます。更に、OS 固有のデザインも。"
-    },
-    {
-      "title": "定期的なリリース",
-      "description": "Floorp は Firefox ESR に基づき、４週間に一回の機能更新を行うラピッドリリースを採用するブラウザーです。Firefox より早くアップデートを提供し、セキュリティ問題を迅速に修正しています。"
-    },
-    {
-      "title": "情報収集なし",
-      "description": "Floorp にはユーザーの情報を集める機能は備わっていません。私たちは、全ユーザーの総ダウンロード数とアップデート回数しか知りません！あなたの個人情報を知る必要はありません。"
-    },
-    {
-      "title": "完全なオープンソース",
-      "description": "Floorp のソースコードは完全公開されており、誰でもソースコードを見たり Floorp を構築することができます。公開するのはブラウザー自体だけではありません。ビルド環境もオープンソースです。"
-    }
-  ],
-  "columns": [
-    {
-      "title": "両サイドに配置できるサイドバー",
-      "description": "サイドバーは１つで満足でしょうか？サイドバーでウェブサイトを開きたいと思いませんか？ Floorp にはウェブパネルとブラウザー管理ツールを表示できるサイドバーが組み込まれており、快適なブラウジングが可能です。<br>サクッと横で見たい！忙しいあなたのためのサイドバーです！ブックマークや履歴から、ウェブサイトまで全てをチェックできます！"
-    },
-    {
-      "title": "細部までカスタマイズできるレイアウト",
-      "description": "ツリー型タブを組み込み、垂直タブとして使用、または併用したり。ブックマークバーのカスタマイズなども。痒い所に手が届く、そんなブラウザーです。<br>こだわりにこだわりたい人のためのカスタムCSSや、基本的なテーマなど、カスタマイズ初心者から玄人までの全ての人に最高のカスタマイズをお届けします！"
-    },
-    {
-      "title": "ウェブとユーザーの双方を第一に考えました",
-      "description": "Floorp はユーザーを追跡しません。また、ウェブサイトでの悪意ある追跡のみをブロックします。Floorp を使えば、ウェブサイト側、ユーザー側の双方に利益のあるインターネットを構築できます。<br>広告で生計を立てるウェブサイト運営者と、プライバシーを守りたいユーザーの双方の意見を取り入れ、有害なトラッカーのみをブロックします。より良いインターネットをFloorpと共に探してみませんか？"
-    }
-  ],
-  "latest-article": "最新の記事",
-  "see-more": "もっとみる",
-  "download-floorp": "最新バージョンの Floorp を入手する"
-}
